ensure server test cannot be interrupted by component tests
